tests/regressions/pyproject.toml
[project]
name = "iknaio-tests-nightly"
version = "0.1.0"
description = "Nightly regression tests for GraphSense REST API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.28.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "docker>=7.0.0",
    "testcontainers>=4.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "deltalake>=1.2.1",
    "pyarrow>=16.0.0",
    "boto3>=1.34.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "regression: mark test as regression test (baseline container vs current)",
    "loki_generated: mark test as auto-generated from Loki API logs (many tests, slower)",
    "deltalake: Delta Lake cross-version compatibility tests",
]
asyncio_mode = "auto"

[tool.ruff]
exclude = [".venv", "__pycache__"]
